First commit to project.
functions.jl contains the function for the non-linear opinion dynamics
for an arbitrary number of dimensions. It also contains functions to
calculate the Jacobian and in-place derivative, which are not used right
now but may be useful later. These functions are definitely useful for
bifurcation analysis, but not (as of now) for automatif differentiation.

Included .gitignore for convenience.